Added maven shade dependency. Refactored PlayerLifetimeStatsLambda so that it can accept the playerId parameter.
