+++ src/main/java/org/morts/PlayerLifetimeStatsLambda.java
@@ -1,0 +1,52 @@
+package org.morts;
+
+import com.amazonaws.services.lambda.runtime.Context;
+import com.amazonaws.services.lambda.runtime.RequestHandler;
+import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
+import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
+import org.morts.util.StatCalculatorUtil;
+import org.morts.util.Statline;
+
+import java.sql.*;
+
+public class PlayerLifetimeStatsLambda implements RequestHandler<APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {
+
+    String dbUrl = System.getenv("DB_URL");
+    String dbUser = System.getenv("DB_USER");
+    String dbPassword = System.getenv("DB_PASSWORD");
+
+    @Override
+    public APIGatewayProxyResponseEvent handleRequest(APIGatewayProxyRequestEvent event, Context context) {
+        String playerId = event.getPathParameters().get("playerId");
+
+        if (playerId == null) {
+            return new APIGatewayProxyResponseEvent()
+                    .withStatusCode(400)
+                    .withBody("Missing playerId parameter");
+        }
+
+        Statline statline;
+        try {
+            statline = getPlayerLifetimeStats(Integer.valueOf(playerId));
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
+
+        return new APIGatewayProxyResponseEvent()
+                .withStatusCode(200)
+                .withBody(statline.toString());
+    }
+
+    public PlayerLifetimeStatsLambda() {
+    }
+
+    public Statline getPlayerLifetimeStats(Integer playerId) throws SQLException, ClassNotFoundException {
+        Class.forName("com.mysql.cj.jdbc.Driver");
+        Connection connection = DriverManager.getConnection(this.dbUrl, this.dbUser, this.dbPassword);
+        PreparedStatement preparedStatement = connection.prepareStatement("select * from game g where player_id = " + playerId + ";");
+        ResultSet rs = preparedStatement.executeQuery();
+        return StatCalculatorUtil.computeStatlineFromGameIterator(rs);
+    }
+}
